refactor: extract ClearCacheView for better cache management and navigation
