+++ dauphin/View/Setting/SettingView.swift
@@ -15,9 +15,6 @@ enum SettingsSection: String, CaseIterable {
 
 struct SettingView: View {
   @ObservedObject var viewModel: AuthViewModel
-  @StateObject private var courseViewModel = CourseViewModel()
-  @State private var showingClearCacheAlert = false
-  @State private var cacheCleared = false
   @State private var selectedSection: SettingsSection? = .account
   @Environment(\.horizontalSizeClass) var horizontalSizeClass
 
@@ -33,26 +30,8 @@ struct SettingView: View {
             .tag(SettingsSection.about)
 
           Section("Data Management") {
-            Label {
-              HStack {
-                Text("Clear Cache")
-                Spacer()
-                if cacheCleared {
-                  Image(systemName: "checkmark.circle.fill")
-                    .foregroundColor(.green)
-                    .transition(.scale)
-                }
-              }
-            } icon: {
-              Image(systemName: "trash")
-                .foregroundColor(.red)
-            }
-            .tag(SettingsSection.cache)
-            .contentShape(Rectangle())
-            // .onTapGesture {
-            //   selectedSection = .cache
-            //   showingClearCacheAlert = true
-            // }
+            Label("Clear Cache", systemImage: "trash")
+              .tag(SettingsSection.cache)
           }
         }
         .navigationTitle("Settings")
@@ -64,34 +43,20 @@ struct SettingView: View {
             LibMainView(viewModel: viewModel)
           case .about:
             AboutUsView()
-          case .cache, .none:
+          case .cache:
+            ClearCacheView()
+          case .none:
             VStack(spacing: 20) {
-              Image(systemName: selectedSection == .cache ? "trash" : "gear")
+              Image(systemName: "gear")
                 .font(.system(size: 60))
                 .foregroundColor(.gray)
-              Text(selectedSection == .cache ? "Cache Management" : "Select a setting")
+              Text("Select a setting")
                 .font(.title2)
                 .foregroundColor(.gray)
-              if selectedSection == .cache {
-                Button("Clear Cache") {
-                  showingClearCacheAlert = true
-                }
-                .buttonStyle(.borderedProminent)
-              }
             }
             .frame(maxWidth: .infinity, maxHeight: .infinity)
             .background(Color(UIColor.systemGroupedBackground))
           }
-        }
-        .alert("Clear Cache", isPresented: $showingClearCacheAlert) {
-          Button("Cancel", role: .cancel) {}
-          Button("Clear", role: .destructive) {
-            clearCache()
-          }
-        } message: {
-          Text(
-            "This will remove all cached course data. You'll need to refresh to reload your courses."
-          )
         }
       }
     } else {
@@ -107,55 +72,12 @@ struct SettingView: View {
           }
 
           Section("Data Management") {
-            Button(action: {
-              showingClearCacheAlert = true
-            }) {
-              Label {
-                HStack {
-                  Text("Clear Cache")
-                  Spacer()
-                  if cacheCleared {
-                    Image(systemName: "checkmark.circle.fill")
-                      .foregroundColor(.green)
-                      .transition(.scale)
-                  }
-                }
-              } icon: {
-                Image(systemName: "trash")
-                  .foregroundColor(.red)
-              }
+            NavigationLink(destination: ClearCacheView()) {
+              Label("Clear Cache", systemImage: "trash")
             }
-            .foregroundColor(Color(UIColor.label))
           }
         }
         .navigationTitle("Settings")
-        .alert("Clear Cache", isPresented: $showingClearCacheAlert) {
-          Button("Cancel", role: .cancel) {}
-          Button("Clear", role: .destructive) {
-            clearCache()
-          }
-        } message: {
-          Text(
-            "This will remove all cached course data. You'll need to refresh to reload your courses."
-          )
-        }
-      }
-    }
-  }
-
-  private func clearCache() {
-    // Use CourseViewModel's clearCache method
-    courseViewModel.clearCache()
-
-    // Show success indicator
-    withAnimation {
-      cacheCleared = true
-    }
-
-    // Reset the indicator after 2 seconds
-    DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-      withAnimation {
-        cacheCleared = false
       }
     }
   }
